Added the files necessary for TCP-communication
Two header and source files used to enable the TCP communication using the message defined by the tcp_data_message.proto file.
tcp_io_device is the actual extension to the mem class which is loaded in the main function. It creates an object of the tcp_connection which handles the data-flow between the external environment and AERA in a background thread.