Cart-Pole Simulator
A new section is added to TestMem::eject for "cart-pole" after the "ball" section. Also extra necessary variables are added to the TestMem class.
The added sections are 3 if statements: the first one has an identifier for cart-pole object, the second statement operates if set_force_y_opcode_ command functions. The third one simulates the cart-poles dynamical equations when the relevant object is identified.
The cart-pole variables, including force, theta, and omega, their properties, and the cart-pole object are declared in test_mem.h file.