Generalization of pattern matching
The pattern extractor and the binding map were updated to allow the general matching of two patterns for which the necessary operators (plus and divide) are implemented. If not, it now simply returns false. For this to work, the opcode of the object/ structure needs to be included in the usr_operators/Operators/operators.cpp file and added to the initialized_opcodes vector.