Performance changes and optimization
Changed operation results to be pointers to Atoms to reduce copy load.

Changed return value of evaluate_expression(...) and build_expression_object(...) to pointer types to remove copy load and remove possible issues of a not well-behaved copy-constructor of LocalObject types.

Added build_and_evaluate_expression(...) for runtime optimization including a check for float Atoms which can be evaluated without the need to build complex expressions.