Minor code change to fix bug when intersecting two binding maps
In the 'contains(...)' method called during intersect of two binding maps, when the checked for code object is an object consisting of floats, only the first one was compared to assess whether they are the same. Now it checks all fields of objects that are floats and only returns true when all parts of both sides are the same.