Orientation-analogy code
This commit adds two replicode examples about analogies regarding the object types (rectangles, spheres, and cylinders) based on which the robot hand learns to decide whether it has to rotate before grabbing or not. The HGSanalogyOrientation file is for the 1D replicode version, and the HGSanalogyOrientation.external is needed for running the Webots version.